+++ src/main/java/App/MainLauncher.java
@@ -1,0 +1,11 @@
+package App;
+
+import LogIn.WelcomeController;
+import LogIn.WelcomePresenter;
+
+public class MainLauncher {
+    public static void main(String[] args) {
+        WelcomeController welcomeController = new WelcomeController();
+        welcomeController.launch();
+    }
+}
