+++ src/main/java/DataObjects/UsersController.java
@@ -37,4 +37,29 @@ public class UsersController {
         }
         controller.saveData("Users.json", users, User.class);
     }
+
+    // fix the rest:
+    public User addUser(String firstName, String lastName, String password){
+        int userID = generateUserID();
+        String passwordHash = password;
+        double balance = 0;
+        String fileDirectory = createFileDirectory(firstName, lastName, userID);
+
+        User user = new User(userID, firstName, lastName, passwordHash, balance);
+        usersController.addUser(user);
+
+        return user;
+    }
+
+    private String generatePassword() {
+        return "";
+    }
+
+    private int generateUserID() {
+        return 0;
+    }
+
+    private String createFileDirectory(String firstName, String lastName, int userID){
+        return firstName + lastName + userID;
+    }
 }
