+++ src/main/java/LogIn/LogInPresenter.java
@@ -1,58 +1,18 @@
 package LogIn;
 
-import javax.swing.*;
-import java.awt.*;
+public class LogInPresenter{
+    private final LogInView logInView;
 
-public class LogInPresenter extends JFrame {
+    public LogInPresenter(LogInController controller){
+        this.logInView = new LogInView(controller);
+    }
 
-    public LogInPresenter() {
-        setTitle("Log In");
-        setSize(400, 300);
-        setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-        setLocationRelativeTo(null);
+    public void showView(){
+        logInView.setVisible(true);
+    }
 
-        JPanel loginPanel = new JPanel();
-        loginPanel.setLayout(new GridLayout(5, 2));
-
-        JLabel firstNameLabel = new JLabel("First Name:");
-        JTextField firstNameField = new JTextField();
-        JLabel lastNameLabel = new JLabel("Last Name:");
-        JTextField lastNameField = new JTextField();
-        JLabel userIDLabel = new JLabel("User ID:");
-        JTextField userIDField = new JTextField();
-        JLabel passwordLabel = new JLabel("Password:");
-        JPasswordField passwordField = new JPasswordField();
-        JButton logInButton = new JButton("Log In");
-
-        loginPanel.add(firstNameLabel);
-        loginPanel.add(firstNameField);
-        loginPanel.add(lastNameLabel);
-        loginPanel.add(lastNameField);
-        loginPanel.add(userIDLabel);
-        loginPanel.add(userIDField);
-        loginPanel.add(passwordLabel);
-        loginPanel.add(passwordField);
-        loginPanel.add(new JLabel());
-        loginPanel.add(logInButton);
-
-        add(loginPanel);
-
-        logInButton.addActionListener(e -> {
-            int userID = Integer.parseInt(userIDField.getText());
-            String firstName = firstNameField.getText();
-            String lastName = lastNameField.getText();
-            String password = new String(passwordField.getPassword());
-
-            LogInController logInController = new LogInController();
-
-            logInController.login(userID, firstName, lastName, password);
-
-            if (logInController.getLoggedUser() != null) {
-                new LoggedInPresenter(logInController.getLoggedUser()).setVisible(true);
-                dispose();
-            } else {
-                JOptionPane.showMessageDialog(this, "Invalid username, first name, last name, or password.", "Login Failed", JOptionPane.ERROR_MESSAGE);
-            }
-        });
+    public void disposeView(){
+        logInView.setVisible(false);
+        logInView.dispose();
     }
 }
