+++ src/main/java/LogIn/SignUpController.java
@@ -1,36 +1,30 @@
 package LogIn;
 
-import DataObjects.User;
 import DataObjects.UsersController;
 
 
 public class SignUpController {
+    private SignUpPresenter signUpPresenter = new SignUpPresenter();
+    private LoggedInPresenter loggedInPresenter = new LoggedInPresenter()
     UsersController usersController = new UsersController();
     
     public SignUpController() {
     }
-    
-    public User addUser(String firstName, String lastName, String password){
-        int userID = generateUserID();
-        String passwordHash = password;
-        double balance = 0;
-        String fileDirectory = createFileDirectory(firstName, lastName, userID);
 
-        User user = new User(userID, firstName, lastName, passwordHash, balance); 
-        usersController.addUser(user);
-
-        return user;
+    public void launch(){
+        signUpPresenter.showView();
     }
 
-    private String generatePassword() {
-        return "";
+    public boolean signUpTriggered(String firstName, String lastName, String password){
+        if (!(firstName.isEmpty() || lastName.isEmpty() || password.isEmpty())) {
+            usersController.addUser(firstName, lastName, password);
+            return true;
+        }
+        return false;
     }
 
-    private int generateUserID() {
-        return 0;
-    }
-
-    private String createFileDirectory(String firstName, String lastName, int userID){
-        return firstName + lastName + userID;
+    public void onSignUpSuccess(){
+        signUpPresenter.disposeView();
+        loggedInPresenter.launch();
     }
 }
