+++ src/main/java/LogIn/WelcomePresenter.java
@@ -1,60 +1,18 @@
 package LogIn;
 
-import Views.SignUpPresenter;
+public class WelcomePresenter {
+    private final WelcomeView welcomeView;
 
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class WelcomePresenter extends JFrame {
-
-    public WelcomePresenter() {
-        setTitle("Welcome to Crazy Bank");
-        setSize(400, 300);
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setLocationRelativeTo(null);
-
-        JPanel mainPanel = new JPanel();
-        mainPanel.setLayout(new BorderLayout());
-        JLabel titleLabel = new JLabel("Welcome to Crazy Bank!", SwingConstants.CENTER);
-        titleLabel.setFont(new Font("Arial", Font.BOLD, 24));
-        JButton loginButton = new JButton("Log In");
-        JButton signUpButton = new JButton("Sign Up");
-
-        JPanel buttonPanel = new JPanel();
-        buttonPanel.setLayout(new FlowLayout());
-        buttonPanel.add(loginButton);
-        buttonPanel.add(signUpButton);
-
-        mainPanel.add(titleLabel, BorderLayout.NORTH);
-        mainPanel.add(buttonPanel, BorderLayout.CENTER);
-
-        add(mainPanel);
-
-        loginButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                new LogInPresenter().setVisible(true);
-                dispose();
-            }
-        });
-
-        signUpButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                new SignUpPresenter().setVisible(true);
-                dispose();
-            }
-        });
+    public WelcomePresenter(WelcomeController controller) {
+        this.welcomeView = new WelcomeView(controller);
     }
 
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-                new WelcomePresenter().setVisible(true);
-            }
-        });
+    public void showView(){
+        welcomeView.setVisible(true);
+    }
+
+    public void disposeView(){
+        welcomeView.setVisible(false);
+        welcomeView.dispose();
     }
 }
