+++ src/main/java/LogIn/WelcomeView.java
@@ -1,0 +1,47 @@
+package LogIn;
+
+import javax.swing.*;
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class WelcomeView extends JFrame {
+
+    public WelcomeView(WelcomeController controller) {
+        setTitle("Welcome to Crazy Bank");
+        setSize(400, 300);
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setLocationRelativeTo(null);
+
+        JPanel mainPanel = new JPanel();
+        mainPanel.setLayout(new BorderLayout());
+        JLabel titleLabel = new JLabel("Welcome to Crazy Bank!", SwingConstants.CENTER);
+        titleLabel.setFont(new Font("Arial", Font.BOLD, 24));
+        JButton loginButton = new JButton("Log In");
+        JButton signUpButton = new JButton("Sign Up");
+
+        JPanel buttonPanel = new JPanel();
+        buttonPanel.setLayout(new FlowLayout());
+        buttonPanel.add(loginButton);
+        buttonPanel.add(signUpButton);
+
+        mainPanel.add(titleLabel, BorderLayout.NORTH);
+        mainPanel.add(buttonPanel, BorderLayout.CENTER);
+
+        add(mainPanel);
+
+        loginButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                controller.logInTriggered();
+            }
+        });
+
+        signUpButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                controller.signUpTriggered();
+            }
+        });
+    }
+}
